requirements_enterprise.txt
Flask==2.3.3
Werkzeug==2.3.7
cryptography==41.0.4
pyotp==2.9.0
qrcode==7.4.2
Pillow==10.0.1
python-magic==0.4.27
psutil==5.9.5
schedule==1.2.0
aiofiles==23.2.1
flask-wtf==1.1.1
wtforms==3.0.1
redis==4.6.0
celery==5.3.1
elasticsearch==8.9.0
gunicorn==21.2.0
supervisor==4.2.5
python-dotenv==1.0.0